Refactor message.ts into separate files for message guards, queries, schema, and internals for chat message management
